requirements.txt
streamlit
pillow
numpy
tensorflow
requests